requirements.txt
Flask>=3.1,<4.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
